+++ maps/game_map_test.go
@@ -98,13 +98,13 @@ func TestMetadataValidate(t *testing.T) {
 
 func TestMapSizes(t *testing.T) {
 	s := FixedSizes(Dimensions{11, 12})
-	require.Equal(t, s[0].Width, uint(11))
-	require.Equal(t, s[0].Height, uint(12))
+	require.Equal(t, s[0].Width, 11)
+	require.Equal(t, s[0].Height, 12)
 
 	s = FixedSizes(Dimensions{11, 11}, Dimensions{19, 25})
 	require.Len(t, s, 2)
-	require.Equal(t, s[1].Width, uint(19))
-	require.Equal(t, s[1].Height, uint(25))
+	require.Equal(t, s[1].Width, 19)
+	require.Equal(t, s[1].Height, 25)
 
 	s = AnySize()
 	require.Len(t, s, 1, "unlimited maps should have just one dimensions")
