+++ maps/registry_test.go
@@ -36,7 +36,7 @@ func TestRegisteredMaps(t *testing.T) {
 			for i := meta.MinPlayers; i < meta.MaxPlayers; i++ {
 				t.Run(fmt.Sprintf("%d players", i), func(t *testing.T) {
 					initialBoardState := rules.NewBoardState(int(mapSize.Width), int(mapSize.Height))
-					for j := uint(0); j < i; j++ {
+					for j := 0; j < i; j++ {
 						initialBoardState.Snakes = append(initialBoardState.Snakes, rules.Snake{ID: fmt.Sprint(j), Body: []rules.Point{}})
 					}
 					err := gameMap.SetupBoard(initialBoardState, testSettings, NewBoardStateEditor(initialBoardState))
@@ -49,7 +49,7 @@ func TestRegisteredMaps(t *testing.T) {
 				for _, mapSize := range meta.BoardSizes {
 					t.Run(fmt.Sprintf("%dx%d map size", mapSize.Width, mapSize.Height), func(t *testing.T) {
 						initialBoardState := rules.NewBoardState(int(mapSize.Width), int(mapSize.Height))
-						for i := uint(0); i < meta.MaxPlayers; i++ {
+						for i := 0; i < meta.MaxPlayers; i++ {
 							initialBoardState.Snakes = append(initialBoardState.Snakes, rules.Snake{ID: fmt.Sprint(i), Body: []rules.Point{}})
 						}
 						err := gameMap.SetupBoard(initialBoardState, testSettings, NewBoardStateEditor(initialBoardState))
